+++ bob-apps/bin/blob_recovery_tool/structs.rs
@@ -82,8 +82,16 @@ impl Validatable for BlobHeader {
 }
 
 impl BlobHeader {
-    pub(crate) fn migrate(mut self, version: u32) -> AnyResult<Self> {
-        self.version = version;
+    pub(crate) fn migrate(self, source_version: u32, target_version: u32) -> AnyResult<Self> {
+        match (source_version, target_version) {
+            (source, target) if source >= target => Ok(self),
+            (0, 1) => self.mirgate_v0_to_v1(),
+            (source, target) => Err(Error::unsupported_migration(source, target).into()),
+        }
+    }
+
+    fn mirgate_v0_to_v1(mut self) -> AnyResult<Self> {
+        self.version = 1;
         Ok(self)
     }
 }
@@ -142,8 +150,8 @@ pub(crate) struct Record {
 }
 
 impl Record {
-    pub(crate) fn migrate(self, source: u32, target: u32) -> AnyResult<Self> {
-        match (source, target) {
+    pub(crate) fn migrate(self, source_version: u32, target_version: u32) -> AnyResult<Self> {
+        match (source_version, target_version) {
             (source, target) if source >= target => Ok(self),
             (0, 1) => self.mirgate_v0_to_v1(),
             (source, target) => Err(Error::unsupported_migration(source, target).into()),